correcao do resultado da divisao na calculadora

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -16,7 +16,7 @@ public class Calculadora {
         int somar = somar(a, b);
         int subtrair = subtrair(a, b);
         int multiplicar = multiplicar(a, b);
-        int dividir = dividir(a, b);
+        double dividir = dividir(a, b);
 
         System.out.println("somar: " + somar);
         System.out.println("subt: " + subtrair);
@@ -36,7 +36,7 @@ public class Calculadora {
         return a * b;
     }
 
-    public static int dividir(int a, int b) {
-        return a / b;
+    public static double dividir(int a, int b) {
+        return (double) a / b;
     }
 }
